+++ docs/src/redis_background_task_async_controller/redis_background_task_async_controller.py
@@ -7,7 +7,7 @@ from apidaora import appdaora, route
 REDIS_URI = f"redis://{os.environ.get('REDIS', '')}"
 
 
-@route.background('/hello', tasks_repository=REDIS_URI)
+@route.background('/hello', tasks_repository_uri=REDIS_URI)
 async def hello_task(name: str) -> str:
     await asyncio.sleep(1)
     return f'Hello {name}!'