+++ docs/src/redis_background_task_lock_args_controller/redis_background_task_lock_args_controller.py
@@ -7,7 +7,9 @@ from apidaora import appdaora, route
 REDIS_URI = f"redis://{os.environ.get('REDIS', '')}"
 
 
-@route.background('/hello-single', lock_args=True, tasks_repository=REDIS_URI)
+@route.background(
+    '/hello-single', lock_args=True, tasks_repository_uri=REDIS_URI
+)
 async def hello_task(name: str) -> str:
     await asyncio.sleep(1)
     return f'Hello {name}!'