+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Improve background task controller
+
 ### 0.23.0 - 2020-04-12
 
  - Create background tasks middlewares